refactor rename blueprintTypes -> installableTypes
