+++ src/extensionConsole/pages/blueprints/gridView/GridView.tsx
@@ -18,7 +18,7 @@
 import styles from "./GridView.module.scss";
 
 import React, { useCallback, useEffect, useMemo, useState } from "react";
-import { type InstallableViewItem } from "@/installables/blueprintsTypes";
+import { type InstallableViewItem } from "@/installables/installableTypes";
 import { VariableSizeList as List } from "react-window";
 import GridCard from "./GridCard";
 import { type Row } from "react-table";
@@ -26,7 +26,7 @@ import ListGroupHeader from "@/extensionConsole/pages/blueprints/listView/ListGr
 import { uuidv4 } from "@/types/helpers";
 import { getUniqueId } from "@/installables/installableUtils";
 import GridCardErrorBoundary from "@/extensionConsole/pages/blueprints/gridView/GridCardErrorBoundary";
-import { BlueprintsPageContentProps } from "@/extensionConsole/pages/blueprints/BlueprintsPageContent";
+import { type BlueprintsPageContentProps } from "@/extensionConsole/pages/blueprints/BlueprintsPageContent";
 
 /**
  *  Expands `react-table` rows recursively in chunks of
